requirements.txt
Django==2.0.1
django-guardian==1.4.9
django-rest-auth==0.9.2
djangorestframework==3.7.7
Pillow==5.0.0
psycopg2==2.7.3.2
pytz==2017.3
six==1.11.0
